requirements.txt
streamlit>=1.30.0
openai>=1.12.0
sentence-transformers>=2.3.0
numpy>=1.24.0
praw>=7.7.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
torch>=2.1.0
